requirements.txt
aiohttp==3.7.3
anyio==2.0.2
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
cryptography==3.4.7
cycler==0.10.0
dataframe-image==0.1.1
decorator==4.4.2
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-seed==0.2.2
djangorestframework==3.12.4
entrypoints==0.3
Faker==8.11.0
httplib2==0.19.1
idna==3.2
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
jedi==0.18.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-server==1.2.1
jupyterlab==3.0.3
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.1
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lab==6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multidict==5.1.0
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
simplejson==3.17.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.1
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
txt2tags==3.7
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.6.3
